1.component/modal責務修正 2.hooks/ファイル名変更 責務修正 3.index hook読み込みファイル修正
